third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-text/white-space/trailing-space-position-001-expected.txt

This is a testharness.js-based test.
[FAIL] CSS Test: Positions of trailing collapsible spaces
  assert_equals: expected 0 but got -0
[FAIL] CSS Test: Positions of trailing collapsible spaces 1
  assert_equals: expected 0 but got -0
[FAIL] CSS Test: Positions of trailing collapsible spaces 2
  assert_equals: expected 0 but got -0
[FAIL] CSS Test: Positions of trailing collapsible spaces 3
  assert_equals: expected 0 but got -0
[FAIL] CSS Test: Positions of trailing collapsible spaces 4
  assert_equals: expected 0 but got -0
Harness: the test ran to completion.
